Docker/usedStuff/src/main/java/pposonggil/usedStuff/service/MemberService.java
package pposonggil.usedStuff.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import pposonggil.usedStuff.domain.Member;
import pposonggil.usedStuff.repository.MemberRepository;

import java.util.List;

@Service
@Transactional(readOnly = true)
@RequiredArgsConstructor
public class MemberService {
    private final MemberRepository memberRepository;

    /**
     * 회원 가입
     */
    @Transactional
    public Long join(Member member) {
        validateDuplicateMember(member);
        memberRepository.save(member);
        return member.getId();
    }

    private void validateDuplicateMember(Member member) {
        List<Member> findNickNames = memberRepository.findByNickName(member.getNickName());
        List<Member> findPhones = memberRepository.findByPhone(member.getPhone());

        if(!findNickNames.isEmpty())
            throw new IllegalStateException("이미 존재하는 닉네임 입니다.");
        if(!findPhones.isEmpty())
            throw new IllegalStateException("이미 존재하는 전화번호 입니다.");

    }

    /**
     * 전체 회원 조회
     */
    public List<Member> findMembers() {
        return memberRepository.findAll();
    }

    /**
     * 회원 아이디로 조회
     */
    public Member findOne(Long memberId) {
        return memberRepository.findOne(memberId);
    }
}
